Refactor: InputTextarea ausgelagert
